+++ modules/transport-grpc/src/main/java/org/opensearch/transport/grpc/proto/request/search/query/QueryBuilderProtoConverterRegistryImpl.java
@@ -48,6 +48,9 @@ public class QueryBuilderProtoConverterRegistryImpl implements QueryBuilderProto
         delegate.registerConverter(new MatchNoneQueryBuilderProtoConverter());
         delegate.registerConverter(new TermQueryBuilderProtoConverter());
         delegate.registerConverter(new TermsQueryBuilderProtoConverter());
+        delegate.registerConverter(new IdsQueryBuilderProtoConverter());
+        delegate.registerConverter(new RangeQueryBuilderProtoConverter());
+        delegate.registerConverter(new TermsSetQueryBuilderProtoConverter());
 
         logger.info("Registered {} built-in query converters", delegate.size());
     }